+++ docs/release_notes.rst
@@ -1,6 +1,15 @@
 Release Notes
 =============
 
+v1.1.5
+------
+
+* Added Kubernetes Support
+* Updatd support for docker-compose and ECS
+* Properly split lines for cmd and entrypoint in ECS
+* Added automated deployment to Pypi
+* Bumped docker python to 3.6
+
 v1.1.4
 ------